fix(rendering): resolve bit-shift overflow, brush state, and dialog parenting
Summary: Fixed potential integer overflows in sprite hashing, corrected brush restoration in UI rendering, and fixed dialog parent assignment.

Changes:
- source/rendering/core/sprite_atlas_cache.cpp — cast look components to uint32_t before bit-shifting to prevent overflow
- source/rendering/ui/sprite_icon_renderer.cpp — store brush by value to ensure correct state restoration after fallback drawing
- source/ui/dialogs/find_dialog.cpp — initialize wxDialog with the provided parent instead of global root

+++ source/rendering/core/sprite_atlas_cache.cpp
@@ -107,7 +107,8 @@ TemplateImage* SpriteAtlasCache::getTemplateImage(uint32_t clientID, const Sprit
         if (img->sprite_index != sprite_index) {
             return false;
         }
-        uint32_t lookHash = img->lookHead << 24 | img->lookBody << 16 | img->lookLegs << 8 | img->lookFeet;
+        uint32_t lookHash = static_cast<uint32_t>(img->lookHead) << 24 | static_cast<uint32_t>(img->lookBody) << 16
+            | static_cast<uint32_t>(img->lookLegs) << 8 | static_cast<uint32_t>(img->lookFeet);
         return outfit.getColorHash() == lookHash;
     });
